created controller and configuration https

+++ src/Catalog.Api/Entities/Item.cs
@@ -6,7 +6,7 @@ namespace Catalog.Api.Entities
     {
         public Guid Id { get; init; }
         public string Name { get; init; }
-        public decimal Price { get; init; }
+        public double Price { get; init; }
         public DateTimeOffset CreatedDate { get; init; }
     }
 }
